+++ semana9/projeto-astromatch/package.json
@@ -7,7 +7,9 @@
     "@testing-library/react": "^11.1.1",
     "@testing-library/user-event": "^12.2.0",
     "axios": "^0.21.0",
+    "rbx": "^2.2.0",
     "react": "^17.0.1",
+    "react-app-rewired": "^2.1.6",
     "react-dom": "^17.0.1",
     "react-lottie": "^1.2.3",
     "react-scripts": "4.0.0",
