+++ semana9/projeto-astromatch/src/App.css
@@ -1,6 +1,6 @@
 .App {
   text-align: center;
-  background-color: #83d6ff62;
+  background-color: #5d0d75da;
   width: 100vw;
   height: 100vh;
   margin: 0;
