+++ semana9/projeto-astromatch/src/components/styled_App.js
@@ -1,0 +1,43 @@
+import styled from "styled-components"
+
+export const AppConteiner = styled.div`
+   background-color: black;
+   height: 100vh;
+   width: 100vw;
+   margin: 0;
+   position: absolute;
+`
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// export const Break = styled.hr`
+//    width: 95%;
+//    height: 1px;
+//    background: linear-gradient(to right, #E783FD, #83D5FF);
+//    margin-bottom: 4%; 
+// `
